Reorganise code
version.ts only has types and Semver now
calculate-version.ts and publish-registry.ts now have the code from
there that was specific to them.
