+++ src/calculate-versions.ts
@@ -1,11 +1,12 @@
-import { FS, getDefinitelyTyped } from "./get-definitely-typed";
-import { Options } from "./lib/common";
-import { CachedNpmInfoClient, UncachedNpmInfoClient } from "./lib/npm-client";
-import { AllPackages } from "./lib/packages";
-import { ChangedPackages, computeAndSaveChangedPackages } from "./lib/versions";
-import { loggerWithErrors, LoggerWithErrors } from "./util/logging";
-import { logUncaughtErrors } from "./util/util";
+import assert = require("assert");
 
+import { FS, getDefinitelyTyped } from "./get-definitely-typed";
+import { Options, writeDataFile } from "./lib/common";
+import { CachedNpmInfoClient, UncachedNpmInfoClient, NpmInfoVersion } from "./lib/npm-client";
+import { AllPackages, TypingsData, NotNeededPackage } from "./lib/packages";
+import { ChangedPackages, ChangedPackagesJson, ChangedTypingJson, Semver, versionsFilename } from "./lib/versions";
+import { loggerWithErrors, LoggerWithErrors } from "./util/logging";
+import { assertDefined, best, logUncaughtErrors, mapDefined, mapDefinedAsync } from "./util/util";
 if (!module.parent) {
     const log = loggerWithErrors()[0];
     logUncaughtErrors(async () => calculateVersions(await getDefinitelyTyped(Options.defaults, log), new UncachedNpmInfoClient(), log));
@@ -18,4 +19,117 @@ export default async function calculateVersions(dt: FS, uncachedClient: Uncached
         const packages = await AllPackages.read(dt);
         return computeAndSaveChangedPackages(packages, log, client);
     });
+}
+
+async function computeAndSaveChangedPackages(
+    allPackages: AllPackages,
+    log: LoggerWithErrors,
+    client: CachedNpmInfoClient): Promise<ChangedPackages> {
+    log.info("Computing changed packages...");
+    const cp = await computeChangedPackages(allPackages, log, client);
+    const json: ChangedPackagesJson = {
+        changedTypings: cp.changedTypings.map(({ pkg: { id }, version, latestVersion }): ChangedTypingJson => ({ id, version, latestVersion })),
+        changedNotNeededPackages: cp.changedNotNeededPackages.map(p => p.name),
+    };
+    await writeDataFile(versionsFilename, json);
+    return cp;
+}
+
+async function computeChangedPackages(
+    allPackages: AllPackages,
+    log: LoggerWithErrors,
+    client: CachedNpmInfoClient): Promise<ChangedPackages> {
+    const changedTypings = await mapDefinedAsync(allPackages.allTypings(), async pkg => {
+        const { version, needsPublish } = await fetchTypesPackageVersionInfo(pkg, client, /*publish*/ true, log);
+        if (needsPublish) {
+            log.info(`Changed: ${pkg.desc}`);
+            const latestVersion = pkg.isLatest ?
+                undefined :
+                (await fetchTypesPackageVersionInfo(allPackages.getLatest(pkg), client, /*publish*/ true)).version;
+            return { pkg, version, latestVersion };
+        }
+        return undefined;
+    });
+    const changedNotNeededPackages = await mapDefinedAsync(allPackages.allNotNeeded(), async pkg => {
+        if (!await isNotNeededPackageAlreadyDeprecated(pkg, client, log)) {
+            log.info(`Now deprecated: ${pkg.name}`);
+            return pkg;
+        }
+        return undefined;
+    });
+    return { changedTypings, changedNotNeededPackages };
+}
+
+
+/** Returns undefined if the package does not exist. */
+interface TypesPackageVersionInfo {
+    readonly version: string;
+    readonly needsPublish: boolean;
+}
+
+async function fetchTypesPackageVersionInfo(
+    pkg: TypingsData,
+    client: CachedNpmInfoClient,
+    canPublish: boolean,
+    log?: LoggerWithErrors): Promise<TypesPackageVersionInfo> {
+    let info = client.getNpmInfoFromCache(pkg.fullEscapedNpmName);
+    let latestVersion = info && getHighestVersionForMajor(info.versions, pkg);
+    let latestVersionInfo = latestVersion && assertDefined(info!.versions.get(latestVersion.versionString));
+    if (!latestVersionInfo || latestVersionInfo.typesPublisherContentHash !== pkg.contentHash) {
+        if (log) { log.info(`Version info not cached for ${pkg.desc}`); }
+        info = await client.fetchAndCacheNpmInfo(pkg.fullEscapedNpmName);
+        latestVersion = info && getHighestVersionForMajor(info.versions, pkg);
+        latestVersionInfo = latestVersion && assertDefined(info!.versions.get(latestVersion.versionString));
+        if (!latestVersionInfo) { return { version: versionString(pkg, /*patch*/ 0), needsPublish: true }; }
+    }
+
+    if (latestVersionInfo.deprecated) {
+        // https://github.com/DefinitelyTyped/DefinitelyTyped/pull/22306
+        assert(
+            pkg.name === "angular-ui-router" || pkg.name === "ui-router-extras",
+            `Package ${pkg.name} has been deprecated, so we shouldn't have parsed it. Was it re-added?`);
+    }
+    const needsPublish = canPublish && pkg.contentHash !== latestVersionInfo.typesPublisherContentHash;
+    const patch = needsPublish ? (latestVersion!.minor === pkg.minor ? latestVersion!.patch + 1 : 0) : latestVersion!.patch;
+    return { version: versionString(pkg, patch), needsPublish };
+}
+
+function versionString(pkg: TypingsData, patch: number): string {
+    return new Semver(pkg.major, pkg.minor, patch).versionString;
+}
+
+async function isNotNeededPackageAlreadyDeprecated(pkg: NotNeededPackage, client: CachedNpmInfoClient, log: LoggerWithErrors): Promise<boolean> {
+    const cachedInfo = client.getNpmInfoFromCache(pkg.fullEscapedNpmName);
+    let latestVersion = cachedInfo && assertDefined(cachedInfo.distTags.get("latest"));
+    let latestVersionInfo = cachedInfo && latestVersion && assertDefined(cachedInfo.versions.get(latestVersion));
+    if (!latestVersionInfo || !latestVersionInfo.deprecated) {
+        log.info(`Version info not cached for ${pkg.desc}`);
+        // Since we're deprecating this package, it should have been published at least once before, so assertDefined.
+        const info = assertDefined(await client.fetchAndCacheNpmInfo(pkg.fullEscapedNpmName));
+        latestVersion = assertDefined(info.distTags.get("latest"));
+        latestVersionInfo = assertDefined(info.versions.get(latestVersion));
+    }
+    return !!latestVersionInfo.deprecated;
+}
+
+function getHighestVersionForMajor(versions: ReadonlyMap<string, NpmInfoVersion>, { major, minor }: TypingsData): Semver | undefined {
+    const patch = latestPatchMatchingMajorAndMinor(versions.keys(), major, minor);
+    return patch === undefined ? undefined : new Semver(major, minor, patch);
+}
+
+/** Finds the version with matching major/minor with the latest patch version. */
+function latestPatchMatchingMajorAndMinor(versions: Iterable<string>, newMajor: number, newMinor: number): number | undefined {
+    const versionsWithTypings = mapDefined(versions, v => {
+        const semver = Semver.tryParse(v);
+        if (!semver) {
+            return undefined;
+        }
+        const { major, minor, patch } = semver;
+        return major === newMajor && minor === newMinor ? patch : undefined;
+    });
+    return best(versionsWithTypings, (a, b) => a > b);
+}
+
+export async function getLatestTypingVersion(pkg: TypingsData, client: CachedNpmInfoClient): Promise<string> {
+    return (await fetchTypesPackageVersionInfo(pkg, client, /*publish*/ false)).version;
 }
