+++ src/index.ts
@@ -2,7 +2,7 @@ export { getDefinitelyTyped } from "./get-definitely-typed";
 export { CachedNpmInfoClient, NpmPublishClient, UncachedNpmInfoClient } from "./lib/npm-client";
 export { AllPackages } from "./lib/packages";
 export { clean } from "./clean";
-export { getLatestTypingVersion } from "./lib/versions";
+export { getLatestTypingVersion } from "./calculate-versions";
 export { default as parseDefinitions } from "./parse-definitions";
 
 export { parseNProcesses } from "./tester/test-runner";
